refactor: use one class for all email manager tests
