--- src/test/kotlin/se/uulm/snowballr/backend/mail/CreateVerificationLinkTest.kt
@@ -1,35 +1,0 @@
-package se.uulm.snowballr.backend.mail
-
-import io.mockk.every
-import io.mockk.mockk
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.BeforeEach
-import org.junit.jupiter.api.Test
-import se.uulm.snowballr.backend.env.Env
-
-class CreateVerificationLinkTest : EmailManagerTest() {
-    private lateinit var emailManager: IEmailManager
-
-    @BeforeEach
-    fun setUp() {
-        val miscellaneousMock = mockk<Env.Miscellaneous>()
-        every { miscellaneousMock.frontendBaseUrl } returns testFrontendUrl
-
-        val envMock = mockk<Env>()
-        every { envMock.miscellaneous } returns miscellaneousMock
-
-        every { envReaderMock.env } returns envMock
-
-        emailManager = EmailManager(envReaderMock, mailerMock, emailTemplateManagerMock)
-    }
-
-    @Test
-    fun `When given a token, then the correct verification link is created`() {
-        val verificationToken = "random-test-token"
-        val expectedLink = "$testFrontendUrl/verifyemail?token=$verificationToken"
-
-        val actualLink = emailManager.createVerificationLink(verificationToken)
-
-        assertThat(actualLink).isEqualTo(expectedLink)
-    }
-}
